+++ History.md
@@ -1,3 +1,7 @@
+* HeathS: WIXBUG:4420 - Suppress patch sequence data to optimize patch performance
+
+* HeathS: WIXBUG:4366 - Correctly enum all products for non-specific patches
+
 * BobArnson: WIXBUG:4443 - Ensure that MsiPackages in a Bundle have ProductVersions that fit in a QWORD, how Burn represents a four-field version number with each field a WORD.
 
 * BobArnson: WIXBUG:3838 - Since the compiler coerces null values to empty strings, check for those in ColumnDefinition.